Extract LayoutIds from roc_gen to roc_mono::layout
