+++ cli/src/repl.rs
@@ -296,7 +296,7 @@ fn gen(src: &[u8], target: Triple, opt_level: OptLevel) -> Result<ReplOutput, Fa
             exposed_to_host: MutSet::default(),
         };
 
-        let mut layout_ids = roc_gen::layout_id::LayoutIds::default();
+        let mut layout_ids = roc_mono::layout::LayoutIds::default();
         let mut headers = Vec::with_capacity(procedures.len());
 
         // Add all the Proc headers to the module.
