+++ compiler/gen/src/lib.rs
@@ -11,7 +11,6 @@
 // re-enable this when working on performance optimizations than have it block PRs.
 #![allow(clippy::large_enum_variant)]
 
-pub mod layout_id;
 pub mod llvm;
 
 pub mod run_roc;
